requirements.txt
fastapi
uvicorn[standard]
selenium
pydantic
python-dotenv